Implement partial parameter dependencies
The plugin is now able to tell when only a part of a parameter is
EXECed, and will backpropagate taintedness to that part only. So for
instance:

  echo $param['x'];

will only set the EXEC bit for the offset "x" in the FunctionTaintedness
object for the corresponding parameter.

Results on MW core:
- No relevant time difference
- +131MB (ouch)
- +5 net issues:
-- +2 copies of an existing issue (bug fixed with
Iae4f6e06e85ce1f9597bd8ff68edb0a7800f0d96)
-- +3 now-unused suppressions
- One issue has different caused-by lines

Bug: T269903
Change-Id: I368e3a6fdcc1c4453fd4c034e29a2c455f8ad23b