Store per-offset method links
The eventual goal is that if we have code like

  $someVar['foo'] = $funcParam;

the links from $funcParam are now stored only for the offset 'foo', and
not for the whole $someVar object. Thus, echoing e.g. $someVar['bar']
will not backpropagate the taintedness to the function.

In practice, this still doesn't work for assignments like the above, but
only when the array is created as a literal, e.g.

  $someVar = [ 'foo' => $funcParam ];

The assignment case is not too complicated, but will be tackled in
another commit. Also, since backpropagation doesn't happen if the
variable is known to have a safe union type, this avoids some false
positives in the assignment case (e.g. echoSafeIndirect2 in the
"dimlinks" test).

It also doesn't work when the offset-get happens directly in the echo
statement (i.e. `echo $x['foo']`, as opposed to `$y = $x['foo']; echo
$y;`), which is yet another not-so-hard-to-fix issue which will be
resolved later.

Note that we still can't handle partial links on the parameter, e.g. in

  $someVar = $funcParam['foo'];

$someVar will be linked to the whole $funcParam. Thus, if $someVar is
later echoed and you call that function with argument

  [ 'foo' => 'this is what is echoed', 'bar' => $_GET['unsafe'] ]

this will still be reported as an XSS. Fixing this is the next step.

All in all, what this patch does is adding the boilerplate needed to
deal with offsets (partly copied from Taintedness), while using that
boilerplate is left for later.

For MW core, compared to the parent patch, we have no difference in
terms of times or emitted issues, and +54 MB.

Change-Id: I821e58eefadd18dedf532fdc4132d2bc573615ae