Implement offset-wise backpropagation for assignments and dims
In assignments, any offset at the LHS is now properly used for tracking
dependencies:

  $x['foo'] = $parameter;

will only add a link to the offset "foo" of $x, not to the whole object.

Additionally, only the relevant links are now retrieved for all DIM
nodes, i.e. if we have:

  $x = $y['foo'];

we now only take the links in the "foo" offset at the RHS.

On MW core, we have:
- +9 MB
- +~5 seconds
- 1 suppression is no longer unnecessary
- Some caused-by lines change

Change-Id: Ia4454ecec2278d2ad6b88bbdc21ef86fe91d5183