Return overhaul 1: use the taintedness of the expression as OVERALL
The previous code was putting the taintedness of a parameter being
returned inside the FunctionTaintedness object, and then intersecting
that with the taintedness of the argument. However, this isn't correct,
because the taintedness of a parameter object doesn't represent which
taintedness is passed through, but rather what taintedness is added by
the function call, which is what the "overall" key is for.

Additionally, remove PRESERVE from overall (as there's no parameter it
can refer to), and keep only PRESERVE for param taintedness.

It's worth noting that in principle, this approach wouldn't work for
code such as

  return htmlspecialchars( $param );

where we wouldn't account for HTML_TAINT being removed. However:
- We only set PRESERVE if the returned expression has a link to a
  parameter
- Function calls do not preserve any link (yet)

which means that the code above won't set a PRESERVE at all.

This will be improved in subsequent commits.

The effects of this patch on MW core are not small:
- 2m 12s -> 2m 8s
- -28 MB
- -42 issues overall, with some new issues being introduced (some are
  unused suppressions)

Change-Id: I5a1c8884e8184cd5e6544eca1b245c1fd7d5d531