Split param taintedness into sink/preserve
The "sink" part only contains EXEC flags, and is used to determine
whether a function call is dangerous.
The "preserve" part only contains YES flags, and is used to determine
the taintedness of the return value (in particular, how the taintedness
of each argument contributes to the taintedness of the return).

This patch also fixes some confusion and mishandling of what taint
should be used e.g. for caused-by lines. As such, some caused-by lines
are now improved, and some get worse; those that got worse were mainly
added just by chance, and not intentionally.

Effects on MW core:
- No significant time difference (perhaps a tad faster)
- -18 MB
- same number of issues, but some caused-by lines changed, as expected.

Change-Id: I46dddcd7e031a04c00163c31a525c9c261de5e98