Return overhaul 2: Use method links to determine what taint is preserved
We now use the method links of the returned expression to determine
which taint is preserved for each parameter. This allows us to remove
MatchReturn visitor.

This change brings improved accuracy for functions that do preserve an
argument, but alter its shape before returning it.

Next steps are selectively preserving some flags but not others, and
then returning method links from a method call, so we can handle nested
calls and the like.

Effects on MW core:
- At most 1 second slower
- + 17 MB
- + 2 issues

Change-Id: I48cbef10329249f506d78abddffec823b39e7fd8