Clean up handleMethodCall
This method is called very often (obviously), so it's important to keep
it easy to understand and refactor. This patch moves some code around so
we can reuse some variables and avoid unnecessary checks (e.g. any
check, if the current parameter is ARRAY_OK). It also adds a parameter
to avoid computing which taintedness is preserved by a function call,
which is used by callers that weren't using the return value. Since
computing this preserve information is going to become more expensive,
it's important that we do that only when really needed.

Change-Id: I9db6247494c5ad2e38bebda69e2754f2fd3b889d