Minor cleanup and optimisation
This one doesn't really do much for performance (saves roughly 1 second
on MW core), but removes some unnecessary stuff.

Most notably, remove the part that would in theory emit an issue for
"always unsafe functions". These are supposed to be functions with EXEC
flags in their overall, but the plugin doesn't do that internally, and
it would also have unclear semantic (as it doesn't say what's being
EXECed).

Change-Id: I49811bcaf02d86916025bb3308c74d1509e2ea15