Special-case and improve handling of Database::insert
First, remove NUMKEY as that's not the correct type. With NUMKEY here,
we would not report tainted array keys (= DB field names) when the
function was called with a single row. And we would consider all the
column values unsafe in the multiple row case. Instead, just use plain
SQL taintedness, but apply it to the keys only.

Second, add a hook to support the multiple rows case, adding the same
SQL taintedness to the keys of each element in $rows.

Bug: T290563
Change-Id: Icd9513f3a5b2b84fbbc805f7d7793f7199df3d14