Avoid calling getTaintOfFunction when unnecessary
The function tries very hard to retrieve a value (e.g. checks alternate
func definitions, analyzes the func body etc.), so it should be used
sparingly. Additionally, it also causes the function taintedness to be
modified, thus confusing things even more. The part of the method that
computed hardcoded/annotated taint was factored out and is now used
when visiting the func declaration in preorder, so we don't have to
worry about it in postorder.

Then, when visiting the declaration in postorder, we now only need to
set an empty taint if none was set.

Consequently, we don't need to use getTaintOfFunction anymore in the
merge helpers, since the annotated taint is guaranteed to be read at
some point. Previously, the code was implicitly relying on those to
be computed when merging taintedness.

Change-Id: I40219ec89dde75914cedb675134c8970cd354a34