Simplify/optimize setting function caused-by
We can do it all in a single method. First, it's shorter and easier to
read. Second, we now manipulate a single FunctionCausedByLines object
and set it on the function at the end of the method, rather than making
a clone of the current object, adding stuff to it, and then setting it
back on the method for each parameter.

Change-Id: Ibc6a735c37c404f9bed1a6430c8b97a5dcead01d