Use actual sink taint for caused-by messages
Instead of using all sink flags that the sink function/construct has,
only use their intersection with what the argument has. This is as
computed by Taintedness::intersectForSink. We do that in maybeEmitIssue,
since this way we don't need to call intersectForSink twice, and
CausedByLines::toStringForIssue is now called in a single place.

Change-Id: If00d66705115a6ea2a5717e971cd20059fbaf367