Don't use Taintedness in preserved param lines
Links are already enough, so no need to convert PreservedTaintedness to
Taintedness. Using PreservedTaintedness instead of MethodLinks would
probably make sense (especially since we already have the right
PreservedTaintedness available in the FunctionTaintedness object), but
currently this would be inconsistent with non-function causedby lines
(which use MethodLinks), so leave that for a future change.

Change-Id: I2f678a041e5b6a0de81a0f70405e689f8ff71c57