Remove $msgArgsGetter hack
This was a performance optimization introduced when retrieving the
caused-by lines was expensive. Nowadays it's very cheap (just calling a
getter), so it doesn't make sense to use a closure for that.

Converting a CausedByLines object to a string for use in issue is still
not trivially cheap, but that's done at the end of maybeEmitIssue, when
we already know that the lines are necessary.

Change-Id: I6fb19016304db56c54992c52d0b367b4f26b453e