Remove INAPPLICABLE_TAINT
Creating taintedness objects for inapplicable nodes has the only effect
of wasting time and memory, since we should never need to get their
taintedness. So optimize that case with a class constant, and ensure
consistency with two structure tests.

Also, use SAFE instead of INAPPLICABLE for nodes that can actually be
used as expression, and for syntax errors.

Effect of this change on MW core:
- Time: 1m48s -> 1m46s
- Memory: 3185MB -> 3171MB

Change-Id: I782f94557bbad958ed7cfda72c232627f2394047