Remove redundant "own" property in ParamLinksOffsets
We don't need it, as we can just check whether the flags are 0. To make
that work, unset flags when pushing an offset, moving them to the offset
being added.

On MW core the runtime is the same, the memory usage decreases from 3233
to 3218 MB, and the code should be easier to read, too.

Change-Id: I928ada3a0d3f7c9dda37d12452c2c731d97c1f59