Split the loop in handleMethodCall into two smaller loops
The first one computes up-to-date argument data, backpropagates
taintedness, and checks if the callee is a sink.

The second loop, which runs only if needed, computes the taintedness
preserved by the function call.

By separating the loops, we no longer need @phan-var for the
previously-complex control flow.

This will also make it easier to special-case some functions and compute
their preserved taintedness in one go, instead of per-argument.

Change-Id: I70b2030800be7372053a9243a52094e0af4f31c7