Add new method for special-casing the preserved taintedness of functions
This is similar to the old one, but is only called once per call, and
not once per argument. Thus, it makes it easier to infer complex shape
mutations.

This patch tries not to change (=improve) the inference too much, except
for cases where it's very easy to do so. Another patch will improve the
inference for the remaining functions.

Change-Id: Id659c5a864ee92f9755b8b0eb1fec7bfa4bcde38