Improve shape inference for array diffs and intersections
This is not perfect, but still much better than the previous code where
we were just preserving all the taintedness from every argument.

Improving the current version is much more difficult because we would
need to know the values inside the arrays, but that's not something we
store in Taintedness, and it would be impractical to do so.

Change-Id: I9880ea6089b9922c52a06ab6704511edbff323f8