Use a "fake" plugin for integration tests
We would like to remove hardcoded taintedness data from the MediaWiki
plugin. However, that's currently being used in a lot of tests. Create a
separate test plugin with only the method information that is actually
needed in tests, so that we can remove it from the real plugin.

Ideally, the long-term plan should be:
- Integration tests that test MW-specific stuff except for
  getCustomFuncTaints (so things like MWVisitor) should just use the
  real MW plugin.
- Tests that cover plugin functionality in general (e.g., providing
  hardcoded information via getCustomFuncTaints) should use a test
  plugin that extends the generic security check plugin.
- Tests that cover specific methods in MW currently hardcoded in
  getCustomFuncTaints should just be moved to MW core, because method
  annotations should eventually replace getCustomFuncTaints.

Bug: T321806
Change-Id: I0a27600fdb87b14c999b1ab6c824901547f8db3e