Make it possible to return FunctionTaintedness from getCustomFuncTaints
Given that we have *-taint annotations for docblocks,
getCustomFuncTaints should generally only be used for complex parameters
that can't easily be described with annotations. Therefore, it makes
sense to allow FunctionTaintedness here, so that the shape can be
provided explicitly. Supporting shapes in annotations would also be
nice, but that's far more complex and is not going to happen soon.

Switch Database::insert & friends to the new syntax, both to demonstrate
what it would look like, and because a subsequent patch will set a
taintedness shape for the 2nd insert() parameter.

Bug: T290563
Change-Id: I625c5c2e830a3de228054a3e074f34682814fde2