Fix tests expecting wrong behaviour of Database::insert()
Unlike select(), insert() shouldn't really use NUMKEY. If the $rows
argument is an array with the key 0, it's treated as a list of rows. It
doesn't really make sense to pass a non-array in this case, because MW
would crash. And therefore, it's not worth expecting any particular
behaviour from taint-check (safe or unsafe).

Change-Id: I93c4bec907e3cc64e1cd3ccad9475e088c1e9a4e