Inherit method annotations from parents anywhere in the hierarchy
getNonParentAncestorFQSENList only returns immediate ancestors, but it
doesn't iterate the whole hierarchy. Use a while loop instead, with a
safeguard against infinite recursion.

As noted in a code comment, it's not very clear what are the
expectations from *-taint annotations (i.e., when they should be
inherited and when they shouldn't). But since we're already
special-casing interfaces, it makes sense not to stop at the parent
class, and instead go through the whole hierarchy.

A practical consequence of this change is that in MW we would be able to
annotate DbQuoter::addQuotes without also annotating the method
implementations in the various Database subclasses.

Bug: T321806
Change-Id: Ia7a2b78fb44e66f4005951ecfa48a33180c035d6