Improve caused-by lines when backpropagation is involved
Whenever some taintedness is backpropagated, update the caused-by lines
of the involved var-like elements with the taintedness currently being
backpropagated. This way, we no longer need to unconditionally include
every line with potentially relevant links when outputting an issue. And
instead, we can update the taintedness only for the lines that are
actually affected by the taintedness being backpropagated.

Also make sure to consistently use a Taintedness with the proper shape
in markAllDependentVarsYes.

Add tests for a few scenarios where this change has a noticeable effect.
Note that the ones involving a non-trivial shape do not currently work
because the shape is simply ignored when retrieving caused-by lines.
However, after this patch the shape is at least preserved along the full
path from the source to the sink, meaning it will be easier to fix this
bug in the future.

Change-Id: I2f472e56d01f25a5c5eeab47f1f348823f02ff1e