Fix getLinks usage in getRelevantLinksForTaintedness
Have the method return LinksSet directly, so that it can decide when
it's safe to destroy the shape, instead of letting the caller do that
when it might be unsafe.

Also rename the method and update its documentation. It doesn't need to
be "temporary", while it's true that it should only be used when the
shape doesn't matter.

Change-Id: Ia024ad1e9c17e2f672c1a38e9cc55d0e1ab854de