Make MethodLinks immutable
Drop all methods that allow external callers to mutate MethodLinks
objects, and use methods that copy the instance.

Introduce a new ::emptySingleton() method to return a singleton with no
links, so that it doesn't have to be instantiated over and over again
(and remain empty in many instances).

Add a smoke test to detect unreasonably expensive code. Previous
patchsets would hang when run on MediaWiki core due to huge arrays, so
this should make such failure obvious during development.

This patch is part of a series to reduce the memory usage of
taint-check. See benchmark in T281491#10205521.

Bug: T281491
Change-Id: Iea8f8542d1fe2e655301fd2d821bfa0fc9e8e7fb