Use Taintedness singletons whenever possible
Introduce a Taintedness::unknownSingleton() method, similar to
safeSingleton() but for unknown taintedness.

Use singletons in all callers that are not mutating the object directly,
i.e., everything outside the Taintedness class as well as some internal
usages.

This patch is part of a series to reduce the memory usage of
taint-check. See benchmark in T281491#10205521.

Bug: T281491
Change-Id: I59f2db3832c81405f67e6d40a4e3fa2b1302595c