Make CausedByLines immutable
Like with Taintedness and MethodLinks, do not allow external callers to
mutate CausedByLines objects. Introduce `::emptySingleton()` to reuse
the same empty instance whenever possible.

Judging from the test output change (which seems correct), a bug has
been fixed in the process.

This patch is part of a series to reduce the memory usage of
taint-check. See benchmark in T281491#10205521.

Bug: T281491
Change-Id: I54fa28d9987f3d2062acabf5293ebab69b0a962e