Implement basic shape tracking for caused-by lines
Unlike other classes (like Taintedness and MethodLinks), shapes are not
tracked using class members, as that would make it harder to preserve a
consistent ordering when merging multiple objects (see
I349d6f217795402b1c7261b698b4f4b589d44c0a for an attempt).

Instead, the Taintedness and MethodLinks instances inside caused-by
entries are mutated as needed, using an interface similar to the one for
Taintedness/MethodLinks.

Note that for hardcoded built-in functions, we mostly collapse the shape
for now (so that the output remains unchanged). This might be improved
in the future.

This results in noticeable improvements to the accuracy of caused-by
lines, highlighted by the test changes (and the TODOs getting resolved).

Unfortunately, it also has a noticeable negative impact on performance:
when run on MediaWiki core, memory usage increases from 2535MB to 2622MB
(+87 MB); that's the price to pay for better error reporting, at least
for the time being. No difference in time usage.

Change-Id: Ife5055c5bf496021ac5b02e0b7075b06edf35190