Fix handling of variadic parameters
If we find a variadic parameter, e.g. `...$args`, it should be treated
as an array of parameters, not as a parameter itself. So, add
special-casing for variadic parameters in MetodLinks, which allows us to
backpropagate EXEC taints correctly when they're found in the function
body.

No changes when run on MW core.

Change-Id: Ie6b7b31e07286bfeb7a45d79f37b7cc1a9f54406