Improve shape tracking in caused-by lines for calls
Use the shapes of the sink and argument to determine the exact shape of
the caused-by lines for function calls. Previously we were just letting
the whole argument through, resulting in unwanted caused-by lines.

Drop a bunch of methods that are no longer needed and were hacky in the
first place (since they ignored the shape).

No new issues on MW core, memory usage drops from 2596MB to 2587MB.

Bug: T281491
Change-Id: I9a42d597fa54e0a19922767b9add289d267b87fa