Avoid adding empty links to caused-by lines
As an optimization, use `null` when links are empty. Also use a special
string representation for the empty links singleton, so it's easier to
spot in the debug output, and to tell if an empty instance is not the
singleton.

Roughly no effects on MW core, memory usage decreases from 2587MB to
2581MB.

Change-Id: I4c62cbc4b49579bca49aea151a4f268e0a229168