Expand recursion fix to the whole analysis stack
In I9759e7fa88f6408033442e752c37c2874dc0486b, we marked recursive
functions as safe while analyzing them. This patch does the same thing
for every function that is currently being analyzed, regardless of
whether it's the current one.

This allows us to cover scenarios where function A calls function B, and
B calls A (or another function C that in turn calls A, etc.).

In the latter scenario, we would previously not detect the recursion and
just keep analyzing the involved function until reaching the maximum
depth (which is thankfully low enough).

No effects on MW core, just a minimal increase in memory usage (2588MB
-> 2589MB).

Change-Id: If511288b87a385b5a5ecfa1396097d3d57955587