Make backpropagated caused-by lines easier to follow
When we are backpropagating the effect of a sink, put the caused-by
lines of the arguments first and reverse them; then add the current
line, and finally lines from the sink. This results in an ordering
that's easier to follow without having to jump to seemingly unrelated
places all the time.

Take for example the new `causedby-setter-order` test. Previously,
caused-by lines would start from the `echo` line, then proceed backwards
to the setter. With this patch, we go to the setter immediately, and
then follow the value as it is passed around, ending up at the sink
line.

In these tests it might seem a small difference, but in real code its
impact is huge. The setter and the sink might be in different files,
separated by dozens of assignments and whatnot. We now step directly
into those assignments one by one, rather than jumping to a completely
different class whose link with the issue itself may not be obvious at
first.

As a nice side effect, lines for annotated methods are now in the
correct place. Take for example the `dbmultiplerows` test: previously we
would start from the `insert` call, then list the assignments that made
the argument to `insert` unsafe, and only at the very end would we say
that `insert` is unsafe due to its annotations. This information is now
presented immediately after the line of the method call.

No effect on MW core.

Change-Id: I54480ea38a86d3f10cbe29a4ccb75afebc082edf