Avoid processing empty objects to save resources
Add a few emptiness checks to avoid cloning empty objects or adding
empty caused-by lines. Skip processing function arguments when the
parameter has a preserve taint set but it's empty. Do not set function
taintedness and caused-by lines from `return` statements when the
function is already known to be safe by its return type. Special-case
the CausedByLines singleton in `toDebugString` so that it can be more
easily spotted while debugging (and it also becomes obvious whether an
empty object is the singleton or a clone).

On MW core, memory usage drops from 2572MB to 2538MB (-34MB).

Bug: T281491
Change-Id: Iad748e9e4fe5e474269d4d1976ddec1505535c37