Simplify code that sets/merges caused-by lines
Inline trivial methods and merge the objects directly. Previously, we
would read the current value, then modify it, store it, and possibly
repeat the process a few times. Now, we just read the value, make all
the necessary modifications to it, and store it back.

Also, factor out the code that computes the caused-by lines to add, so
we can call it just once and add those lines at different times to
improve ordering (done in subsequent commits).

On MW core, memory usage is slightly increased (2564MB -> 2567MB).

Change-Id: I18ee210734b041ce7645b21dc531efee090a1a79