Improve caused-by accuracy for unsafe method calls
For method calls, caused-by lines should receive a different treatment
depending on whether they refer to the sink (why does the method do
something unsafe) or the argument (why is the argument user-controlled).
Taintedness objects inside the former track what lines do the unsafe
things, and therefore should be treated as a sink taintedness when being
matched to the argument taintedness. On the other hand, the argument
caused-by lines should be matched against the overall taintedness of the
sink.

So, update the relevant methods in CausedByLines to apply the correct
behaviour depending on the source of the lines. For `maybeEmitIssue`,
this additional bit of information is passed by wrapping the
CausedByLines object in an array. This is admittedly not ideal, and
should be refactored someday (which would have to be done anyway when we
replaced "Caused by" with clearer indications in the output, T351366).

Also fix a bug in `appliedToTaintednessForBackprop` that was giving us
the incorrect taintedness in the first place.

No changes on MW core.

Change-Id: Ia9b7776bf47f6d7a9fdd3c931ecda244b775a482