Reorder caused-by lines for variable links
When backpropagating the effect of calling a function with a tainted
argument (most commonly, calling a property setter), we now put the
caused-by lines of the argument before those of the setter. This way,
it's easier to follow where the taintedness is coming from and where
it's going; see the "dependent" test for an example, noticing that we
now put the setter call before the setter body.

No changes on MW core, minuscule difference in memory usage (2561MB
-> 2560MB).

Change-Id: I0336aaa2287802f380e1b0e45a555c81d5f35f4f