Add local variables for links of hardcoded functions
This is a no-op for the time being, but it will eventually become
necessary for passing links through function calls
(I1bd8ae302e91a2b6b951953bc321ea6ae89d5955). Since the method is very
long, this change adds local variables for links, so as to have a
smaller diff for the real change later.

No changes on MW core, as expected.

Change-Id: Ia8e044ff31e93b63abdc8bf8e79fe31615d584d7