Improve caused-by lines for assignments
Add a new method to append a line for an assignment, which never
modifies existing method links. This is mostly a no-op for now, but it
will be needed by later patches. This approach is also quite hacky, but
the whole CausedByLines business is, specifically the merging, and I'm
hoping to rewrite them one day.

No new issues on MW core, memory usage slightly decreased (2523MB ->
2521MB).

Change-Id: I7b5b5d2620207334d8d856b68755ad18d920fe07