Fix bug in preserved caused-by lines for pass-by-ref
Shockingly, `asPreservedForArgument` must be used on the argument lines,
and `asPreservedForParameter` on the parameter lines. Who would've
thought. So, actually do that instead of treating all lines as parameter
lines.

I don't have a testcase for this, because everything I tried involved
links being preserved by a function call, which is only going to be
implemented in I1bd8ae302e91a2b6b951953bc321ea6ae89d5955. On the other
hand, some of our existing test would fail with that patch applied and
without this change, so that should be sufficient.

On MW core, memory usage is reduced (2536MB -> 2523MB); no new issues,
and speed is roughly the same.

Bug: T281491
Change-Id: I86b1c578538e14e9a09dd38b37731e0c81430313