Batch shape updates when visiting arrays
Instead of updating the shapes one element at a time, do all updates at
the end, when we have the entire shape, so we don't need to create/clone
a ton of intermediate objects that are then immediately discarded.

This drastically improves performance when analyzing large arrays. The
test for that now uses an array with 100k elements (10x), and completes
in 1 second, whereas it would previously take a few minutes.

On MW core: no new issues, tiny increase in memory usage (2607MB ->
2609MB), and possibly a second or two faster.

Bug: T395292
Bug: T359003
Change-Id: I33a16d21280445f975eac175445526824c4f78d0