Update for new Message FQSEN in MediaWiki core
\Message is now a deprecated alias and has been for a while, see
I088cbc53fbcdb974e5b05b45a62e91709dacc024.

In MediaWikiSecurityCheckPlugin, update the logic that selectively
disables DoubleEscaped checks to use both FQSENs.

In tests, drop everything that is really just testing the Message class;
tests for that are now in MW core's TaintCheckAnnotationsTest.
Generalize a couple tests and repurpose them to explicitly test
hardcoded taintedness. And finally, duplicate the test verifying the
DoubleEscaped logic, so that it tries both the old alias and the new
FQSEN.

Bug: T353458
Change-Id: I70dfd3ebce79febfa134bb3b2ebd33dcd5851462