Fix numkey taint detection for SQLi taint sinks
If the RHS's overall flags have SQL taint, that means in particular that
the keys could have it, as we weren't able to infer the shape. So, add
NUMKEY in that case, as this allows detection of some SQLi issues.

This remains hacky, as we should check the union type shape, not
taintedness, but this is all we can do for the time being.

Also fix a test case that thought it was passing $options, but really it
was passing $join_conds. Both tests were labeled as "unsafe" but never
emitted any issues.

Change-Id: Iddd4a5706ac189bbfa375441a12b7d973bd8c6a0