Update supported hook handler styles
Drop support for legacy style deprecated in MW 1.41 and to be dropped
soon (see T401532).

Add support for the remaining array callable case, as well as
first-class callables.

Catch invalid FQSENs where needed, might happen because it's
user-provided code.

There are a few possibilities for cleanup / deduplication here, but
these will be done in another commit.

Bug: T250499
Change-Id: I004504b205f1e06cba21bb7b62a586b7e907a44f