Fix isHTML handling in parser function hooks
Phan's `UnionType::containsTrue` also returns true if the type is just
`false`. Presumably, this is done because `false` in PHPDoc doesn't
prevent the value from actually being true. However, we don't want that
here. So, change the code to manually check for bool and false.

Also switch to using the real type, as that's more consistent with the
previous code analyzing the AST structure.

Follow-up to I614f55e69814528.

Change-Id: Ib4fbafe36fb9bbba8331a9d616cfe1814f8a052f