Improve error reporting for numkey taint
Instead of adding flags to the sink that might not actually be there, do
the normalization the other way around, adding numkey to SQL in the
expression taint. This way, the `withOnly` does not wipe the entire
shape (numkey is not included in YES_TAINT) and it can later be
processed in `intersectForSink`.

Bug: T401748
Change-Id: I65ce47727a94bedf6018169b3f3ab506bfc56813